Add support for manual retry.
Move failed downloads to the Finished list, and present an action sheet
for them.
Add a copy constructor for SDSafariDownload, as we can't un-finish an
NSOperation.
Retry is achieved by making a clone of the download to retry and
enqueueing it.
Don't display progress bars on failed downloads.
Don't try to resume a download if we can't find its tempfile.